LayoutTests/platform/mac/accessibility/progressbar-expected.txt

CONSOLE MESSAGE: line 40: TypeError: Result of expression 'document.getElementById("progressbar3")' [null] is not an object.
X X
This test makes sure that progress element can be accessed by Accessibility FW.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.intValue is 7
PASS obj.intValue is 0
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
